Split message module into submodules
