+++ src/lotus/mod.rs
@@ -1,22 +1,22 @@
-pub mod client;
 ///! The lotus api to interact with lotus node
-pub mod message;
-#[cfg(test)]
-mod tests;
-
 use std::collections::HashMap;
+use std::fmt::Debug;
+
 use anyhow::Result;
 use async_trait::async_trait;
 use cid::Cid;
 use fvm_shared::address::Address;
 use serde::de::DeserializeOwned;
-use std::fmt::Debug;
 
-pub use crate::lotus::client::LotusJsonRPCClient;
-pub use crate::lotus::message::{MpoolPushMessage, MpoolPushMessageResponseInner};
-pub use crate::lotus::message::{
-    ChainHeadResponse, ReadStateResponse, StateWaitMsgResponse, WalletKeyType, WalletListResponse,
-};
+use message::chain::ChainHeadResponse;
+use message::mpool::{MpoolPushMessage, MpoolPushMessageResponseInner};
+use message::state::{ReadStateResponse, StateWaitMsgResponse};
+use message::wallet::{WalletKeyType, WalletListResponse};
+
+pub mod client;
+pub mod message;
+#[cfg(test)]
+mod tests;
 
 /// The network version of lotus network.
 /// see https://github.com/filecoin-project/go-state-types/blob/f6fd668a32b4b4a0bc39fd69d8a5f8fb11f49461/network/version.go#L7
@@ -41,7 +41,10 @@ pub trait LotusClient {
     async fn state_network_version(&self, tip_sets: Vec<Cid>) -> Result<NetworkVersion>;
 
     /// Returns the CID of the builtin actors manifest for the given network version, see https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v1-unstable-methods.md#stateactormanifestcid
-    async fn state_actor_code_cids(&self, network_version: NetworkVersion) -> Result<HashMap<String, Cid>>;
+    async fn state_actor_code_cids(
+        &self,
+        network_version: NetworkVersion,
+    ) -> Result<HashMap<String, Cid>>;
 
     /// Get the default wallet of the node, see: https://lotus.filecoin.io/reference/lotus/wallet/#walletdefaultaddress
     async fn wallet_default(&self) -> Result<Address>;
